api/requirements/tests.txt
-r ./base.txt

factory-boy==3.2.1
pytest==7.1.2
pytest-django==4.5.2
pytest-sugar==0.9.4
